+++ DesignerRHN/DesignerRHN-ejb/src/java/Entidades/ParametrosTiempos.java
@@ -1,6 +1,7 @@
 package Entidades;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.Date;
 import javax.persistence.*;
@@ -32,9 +33,9 @@ public class ParametrosTiempos implements Serializable {
     @Temporal(TemporalType.TIMESTAMP)
     private Date fechahasta;
     @Column(name = "CODIGOEMPLEADODESDE")
-    private BigInteger codigoempleadodesde;
+    private BigDecimal codigoempleadodesde;
     @Column(name = "CODIGOEMPLEADOHASTA")
-    private BigInteger codigoempleadohasta;
+    private BigDecimal codigoempleadohasta;
     @Column(name = "FECHACONTROLNOVEDAD")
     @Temporal(TemporalType.TIMESTAMP)
     private Date fechacontrolnovedad;
@@ -92,19 +93,19 @@ public class ParametrosTiempos implements Serializable {
         this.fechahasta = fechahasta;
     }
 
-    public BigInteger getCodigoempleadodesde() {
+    public BigDecimal getCodigoempleadodesde() {
         return codigoempleadodesde;
     }
 
-    public void setCodigoempleadodesde(BigInteger codigoempleadodesde) {
+    public void setCodigoempleadodesde(BigDecimal codigoempleadodesde) {
         this.codigoempleadodesde = codigoempleadodesde;
     }
 
-    public BigInteger getCodigoempleadohasta() {
+    public BigDecimal getCodigoempleadohasta() {
         return codigoempleadohasta;
     }
 
-    public void setCodigoempleadohasta(BigInteger codigoempleadohasta) {
+    public void setCodigoempleadohasta(BigDecimal codigoempleadohasta) {
         this.codigoempleadohasta = codigoempleadohasta;
     }
