+++ DesignerRHN/DesignerRHN-ejb/src/java/InterfacePersistencia/PersistenciaTurnosEmpleadosInterface.java
@@ -6,6 +6,7 @@
 package InterfacePersistencia;
 
 import Entidades.TurnosEmpleados;
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.Date;
 import java.util.List;
@@ -21,9 +22,9 @@ public interface PersistenciaTurnosEmpleadosInterface {
 
     public Date obtenerFechaFinalMaximaTurnosEmpleados(EntityManager em);
 
-    public int ejecutarPKG_CONTARNOVEDADESLIQ(EntityManager em, Date fechaDesde, Date fechaHasta, BigInteger emplDesde, BigInteger emplHasta);
+    public int ejecutarPKG_CONTARNOVEDADESLIQ(EntityManager em, Date fechaDesde, Date fechaHasta, BigDecimal emplDesde, BigDecimal emplHasta);
 
-    public void ejecutarPKG_ELIMINARLIQUIDACION(EntityManager em, Date fechaDesde, Date fechaHasta, BigInteger emplDesde, BigInteger emplHasta);
+    public void ejecutarPKG_ELIMINARLIQUIDACION(EntityManager em, Date fechaDesde, Date fechaHasta, BigDecimal emplDesde, BigDecimal emplHasta);
 
     public void crear(EntityManager em, TurnosEmpleados turnosEmpleados);
